refactor: tc 분리
