+++ src/main/java/step3/RacingCondition.java
@@ -21,11 +21,15 @@ public class RacingCondition {
 
     public static RacingCondition getInput() {
         Scanner scanner = new Scanner(System.in);
-        System.out.println("자동차 대수는 몇 대 인가요?");
-        int carCount = scanner.nextInt();
-        System.out.println("시도할 회수는 몇 회 인가요?");
-        int attemptCount = scanner.nextInt();
+
+        int carCount = getIntInput(scanner, "자동차 대수는 몇 대 인가요?");
+        int attemptCount = getIntInput(scanner, "시도할 회수는 몇 회 인가요?");
 
         return new RacingCondition(carCount, attemptCount);
+    }
+
+    private static int getIntInput(Scanner scanner, String question) {
+        System.out.println(question);
+        return scanner.nextInt();
     }
 }
